requirements.txt
aioredis==2.0.1
alembic==1.8.1
anyio==3.6.1
async-timeout==4.0.2
asyncpg==0.26.0
click==8.1.3
databases==0.6.1
Deprecated==1.2.13
fastapi==0.85.0
greenlet==1.1.3.post0
h11==0.14.0
idna==3.4
Mako==1.2.3
MarkupSafe==2.1.1
packaging==21.3
psycopg2-binary==2.9.4
pydantic==1.10.2
pyparsing==3.0.9
python-dotenv==0.21.0
redis==4.3.4
shell==1.0.1
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
typing_extensions==4.4.0
uvicorn==0.18.3
wrapt==1.14.1

passlib~=1.7.4
PyJWT~=2.6.0
Authlib~=1.1.0
Werkzeug==0.11.15
jose~=1.0.0